+++ src/cyber80/palette_color/type.go
@@ -1,7 +1,7 @@
 package cyber80_palette_color
 
 type Type struct {
-	value int
+	value byte
 	loaded bool
 	errored bool
 }
